+++ mod-source-record-manager-server/src/test/java/org/folio/rest/impl/changeManager/ChangeManagerAPITest.java
@@ -1164,6 +1164,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     verify(1, getRequestedFor(urlEqualTo(LOAN_TYPES_URL)));
     verify(1, getRequestedFor(urlEqualTo(ITEM_NOTE_TYPES_URL)));
     verify(1, getRequestedFor(urlEqualTo(AUTHORITY_NOTE_TYPES_URL)));
+    verify(1, getRequestedFor(urlEqualTo(AUTHORITY_SOURCE_FILES_URL)));
     verify(1, getRequestedFor(urlEqualTo(FIELD_PROTECTION_SETTINGS_URL)));
     verify(1, getRequestedFor(urlEqualTo(TENANT_CONFIGURATIONS_SETTINGS_URL)));
     async.complete();
@@ -1204,6 +1205,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     WireMock.stubFor(get(LOAN_TYPES_URL).willReturn(serverError()));
     WireMock.stubFor(get(ITEM_NOTE_TYPES_URL).willReturn(serverError()));
     WireMock.stubFor(get(AUTHORITY_NOTE_TYPES_URL).willReturn(serverError()));
+    WireMock.stubFor(get(AUTHORITY_SOURCE_FILES_URL).willReturn(serverError()));
     WireMock.stubFor(get(FIELD_PROTECTION_SETTINGS_URL).willReturn(serverError()));
     WireMock.stubFor(get(TENANT_CONFIGURATIONS_SETTINGS_URL).willReturn(serverError()));